20.27.9
+ When entering an area, all enemies that must be fought are displayed
+ Area descriptions can now be procedurally generated
* Reorganixed text generation code
* Recompiled dist
